Move platform specific code to platform directory
