+++ common/wireguard/src/platform/linux/mod.rs
@@ -1,0 +1,1 @@
+pub(crate) mod tun_device;
