+++ common/wireguard/src/platform/mod.rs
@@ -1,0 +1,2 @@
+#[cfg(target_os = "linux")]
+pub(crate) mod linux;
