+++ gateway/Cargo.toml
@@ -62,8 +62,6 @@ nym-statistics-common = { path = "../common/statistics" }
 nym-task = { path = "../common/task" }
 nym-types = { path = "../common/types" }
 nym-validator-client = { path = "../common/client-libs/validator-client" }
-
-[target.'cfg(target_os = "linux")'.dependencies]
 nym-wireguard = { path = "../common/wireguard", optional = true }
 
 [build-dependencies]
